# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * Inventory management API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface MessageResponse {
  message: string;
}

export interface User {
  id: number;
  email: string;
  name: string;
  createdAt: string;
}

export interface RegisterBody {
  email: string;
  password: string;
  name: string;
}

export interface LoginBody {
  email: string;
  password: string;
}

export interface GoogleLoginBody {
  /** @minLength 1 */
  idToken: string;
}

export interface ChangePasswordBody {
  /** @minLength 1 */
  currentPassword: string;
  /** @minLength 8 */
  newPassword: string;
}

export interface AuthResponse {
  user: User;
  message: string;
}

export interface Category {
  id: number;
  name: string;
  /** @nullable */
  description: string | null;
  /** @nullable */
  parentId: number | null;
  userId: number;
  /** @nullable */
  color: string | null;
  /** @nullable */
  icon: string | null;
  createdAt: string;
}

export interface CategoryNode {
  id: number;
  name: string;
  /** @nullable */
  description: string | null;
  /** @nullable */
  parentId: number | null;
  userId: number;
  /** @nullable */
  color: string | null;
  /** @nullable */
  icon: string | null;
  createdAt: string;
  children: CategoryNode[];
  itemCount: number;
}

export interface CreateCategoryBody {
  name: string;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  parentId?: number | null;
  /** @nullable */
  color?: string | null;
  /** @nullable */
  icon?: string | null;
}

export interface UpdateCategoryBody {
  name?: string;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  parentId?: number | null;
  /** @nullable */
  color?: string | null;
  /** @nullable */
  icon?: string | null;
}

export interface Item {
  id: number;
  name: string;
  /** @nullable */
  description: string | null;
  quantity: number;
  /** @nullable */
  photoUrl?: string | null;
  /** @nullable */
  unitId?: number | null;
  /** @nullable */
  unitSymbol?: string | null;
  /** @nullable */
  unitName?: string | null;
  /** @nullable */
  unit: string | null;
  /** @nullable */
  locationId?: number | null;
  /** @nullable */
  locationName?: string | null;
  /** @nullable */
  location: string | null;
  /** @nullable */
  sku: string | null;
  /** @nullable */
  barcode: string | null;
  /** @nullable */
  tags: string | null;
  /** @nullable */
  notes: string | null;
  /** @nullable */
  categoryId: number | null;
  /** @nullable */
  categoryName: string | null;
  userId: number;
  createdAt: string;
  updatedAt: string;
}

export interface CreateItemBody {
  name: string;
  /** @nullable */
  description?: string | null;
  quantity: number;
  /** @nullable */
  photoUrl?: string | null;
  /** @nullable */
  unitId?: number | null;
  /** @nullable */
  unit?: string | null;
  /** @nullable */
  locationId?: number | null;
  /** @nullable */
  location?: string | null;
  /** @nullable */
  sku?: string | null;
  /** @nullable */
  barcode?: string | null;
  /** @nullable */
  tags?: string | null;
  /** @nullable */
  notes?: string | null;
  /** @nullable */
  categoryId?: number | null;
}

export interface UpdateItemBody {
  name?: string;
  /** @nullable */
  description?: string | null;
  quantity?: number;
  /** @nullable */
  photoUrl?: string | null;
  /** @nullable */
  unitId?: number | null;
  /** @nullable */
  unit?: string | null;
  /** @nullable */
  locationId?: number | null;
  /** @nullable */
  location?: string | null;
  /** @nullable */
  sku?: string | null;
  /** @nullable */
  barcode?: string | null;
  /** @nullable */
  tags?: string | null;
  /** @nullable */
  notes?: string | null;
  /** @nullable */
  categoryId?: number | null;
}

export interface DashboardStats {
  totalItems: number;
  totalCategories: number;
  totalQuantity: number;
  lowStockItems: number;
  recentlyAdded: number;
}

export interface CategoryCount {
  categoryId: number;
  categoryName: string;
  itemCount: number;
  /** @nullable */
  color: string | null;
}

export interface ShoppingListItem {
  id: number;
  /** @nullable */
  itemId?: number | null;
  name: string;
  quantity: number;
  /** @nullable */
  unit?: string | null;
  /** @nullable */
  note?: string | null;
  checked: boolean;
  createdAt: string;
}

export interface AddToShoppingListBody {
  itemId?: number;
  name: string;
  quantity?: number;
  unit?: string;
  note?: string;
}

export interface UpdateShoppingListBody {
  checked?: boolean;
  quantity?: number;
  /** @nullable */
  note?: string | null;
}

export interface ForgotPasswordBody {
  email: string;
}

export interface ResetPasswordBody {
  token: string;
  /** @minLength 8 */
  password: string;
}

export type ListItemsParams = {
categoryId?: number;
search?: string;
locationId?: number;
includeSubcategories?: boolean;
};

export type GetRecentItemsParams = {
limit?: number;
};