# lib/api-zod/src/generated/types/resetPasswordBody.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * Inventory management API
 * OpenAPI spec version: 0.1.0
 */

export interface ResetPasswordBody {
  token: string;
  /** @minLength 8 */
  password: string;
}